feat: start of patch downloading

+++ XIVLauncher/Game/Patch/PatchList/PatchListParser.cs
@@ -23,7 +23,7 @@ namespace XIVLauncher.Game.Patch.PatchList
                 
                 output.Add(new PatchListEntry()
                 {
-                    Length = int.Parse(fields[0]),
+                    Length = long.Parse(fields[0]),
                     VersionId = fields[4],
 
                     // bootver patchlists don't have a hash field
